refactor: inline BLS spec vectors as bls.rs child test module
bls_spec_tests.rs was a sibling wired through mod.rs — the odd shape:
it tests bls but wasn't owned by it. With the swept bls.rs at ~16 lines
of code, the 73-line vector harness fits inline as
#[cfg(test)] mod spec_tests (the name keeps the vendored-vectors
signal). Kills the two mod.rs wiring lines; doc references updated.
